Sync patches with server in designer
For security reasons it's important to prevent canvas from having write
access to db.

Here I moved old sync logic to designer folder and got rid of one more
event in pubsub, used atom directly.
