--- apps/designer/app/canvas/shared/sync/index.ts
@@ -1,1 +1,0 @@
-export * from "./use-sync";
