+++ apps/designer/app/shared/sync/index.ts
@@ -1,2 +1,1 @@
-export * from "./sync-status";
 export * from "./sync-stores";
