--- apps/designer/app/shared/sync/sync-status.ts
@@ -1,1 +1,0 @@
-export type SyncStatus = "syncing" | "idle";
